Move astc options into astc_encode.h
